requirements.txt
notebook
pyFAI
xrdfit
fabio
numpy
tqdm
pyyaml